Make "source code" a defined term.
"source code" is only colloquially understood by developers to have a
meaning, and even among developers, there is debate and discussion about
what it should include.  This change proposes a definition for "source
code" for use as part of this legislation.

The definition herein proposed is based in part on the definition found
FSF's GPLv3.  While FOSSA doesn't propose only copyleft licenses by any
means, only copyleft licenses tend to include a definition of what
"source code" means, since such a definition is not central to the legal
functioning of non-copyleft licenses (as it is with copyleft ones).

Since the term "source code" is used extensively in prior definitions,
and given the common confusions even among developers about what is
required, such definition is essential to ensure this legislation
reaches its goals.