Create Limits for Rigging
Limit the circumstances that a person other than a licensed rigger can supervise the hoisting or lowering of articles on the outside of a building. If this rigging is performed by or under the supervision of a person who is not a licensed rigger, such person must complete a Department of Buildings-approved training course.